# dsa40applicationhelper/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DSA40 Application Helper",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/question": {
      "get": {
        "summary": "Get All Dsa Questions",
        "operationId": "/api/question-get_all_dsa_questions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Dsa Question",
        "operationId": "/api/question-add_dsa_question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDSAQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/question/{id}": {
      "get": {
        "summary": "Get Dsa Question",
        "operationId": "/api/question/{id}-get_dsa_question",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions": {
      "get": {
        "summary": "Applicable Questions",
        "operationId": "/api/questions-applicable_questions",
        "parameters": [
          {
            "name": "vlopse",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Vlopse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DSAQuestion"
                  },
                  "title": "Response /Api/Questions-Applicable Questions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform-field-index": {
      "get": {
        "summary": "Platform Field Index",
        "operationId": "/api/platform-field-index-platform_field_index",
        "parameters": [
          {
            "name": "vlopse",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Vlopse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/PlatformFieldSource"
                  },
                  "title": "Response /Api/Platform-Field-Index-Platform Field Index"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validate": {
      "post": {
        "summary": "Validate Answers",
        "operationId": "/api/validate-validate_answers",
        "parameters": [
          {
            "name": "vlopse",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Vlopse"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transform": {
      "post": {
        "summary": "Transform Answers",
        "operationId": "/api/transform-transform_answers",
        "parameters": [
          {
            "name": "vlopse",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Vlopse"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/condition": {
      "get": {
        "summary": "Get Conditions",
        "operationId": "/api/condition-get_conditions",
        "parameters": [
          {
            "name": "vlopse",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Vlopse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schemas/{name}": {
      "get": {
        "summary": "Get Schema",
        "operationId": "/api/schemas/{name}-get_schema",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "vlopse",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Vlopse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Answer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "value"
        ],
        "title": "Answer"
      },
      "AnswerRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "AnswerRequest"
      },
...
      "DSAQuestion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          },
          "input_type": {
            "$ref": "#/components/schemas/InputType"
          },
          "help_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help Text"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "granularity": {
            "type": "string",
            "title": "Granularity",
            "default": "general"
          },
          "source_general_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source General Id"
          },
          "vlopse": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlopse"
          },
          "group_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "input_type"
        ],
        "title": "DSAQuestion"
      },
      "ErrorDetails": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "type_error",
              "value_error"
            ],
            "title": "Type"
          },
          "loc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Loc"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "type",
          "loc",
          "message"
        ],
        "title": "ErrorDetails"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InputType": {
        "type": "string",
        "enum": [
          "text",
          "file_upload",
          "date_select",
          "selection",
          "multi_select",
          "iso-3166-1",
          "repeatable_group",
          "composite_group"
        ],
        "title": "InputType"
      },
...
      "PlatformFieldSource": {
        "properties": {
          "vlopse": {
            "type": "string",
            "title": "Vlopse"
          },
          "platform_id": {
            "type": "string",
            "title": "Platform Id"
          },
          "platform_text": {
            "type": "string",
            "title": "Platform Text"
          },
          "general_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "General Ids"
          },
          "helper_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Helper Labels"
          },
          "helper_section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Helper Section"
          }
        },
        "type": "object",
        "required": [
          "vlopse",
          "platform_id",
          "platform_text",
          "general_ids",
          "helper_labels"
        ],
        "title": "PlatformFieldSource"
      },
...
      "PostDSAQuestionRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "input_type": {
            "$ref": "#/components/schemas/InputType"
          },
          "help_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help Text"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "input_type"
        ],
        "title": "PostDSAQuestionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationResponse": {
        "properties": {
          "errors": {
            "additionalProperties": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ErrorDetails"
                  },
                  "type": "array"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Errors"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "kind": {
            "type": "string",
            "enum": [
              "validation",
              "transformation"
            ],
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "errors",
          "ok",
          "kind"
        ],
        "title": "ValidationResponse"
      },
...
    }
  }
}